add load assessment helper class
