+++ src/DetailsView/components/load-assessment-button.tsx
@@ -5,6 +5,7 @@ import * as React from 'react';
 import { InsightsCommandButton } from 'common/components/controls/insights-command-button';
 import { DetailsViewActionMessageCreator } from 'DetailsView/actions/details-view-action-message-creator';
 import { AssessmentDataParser } from 'common/assessment-data-parser';
+import { LoadAssessmentHelper } from 'DetailsView/components/load-assessment-helper';
 
 export type LoadAssessmentButtonDeps = {
     detailsViewActionMessageCreator: DetailsViewActionMessageCreator;
@@ -12,6 +13,7 @@ export type LoadAssessmentButtonDeps = {
 };
 export interface LoadAssessmentButtonProps {
     deps: LoadAssessmentButtonDeps;
+    loadAssessmentHelper: LoadAssessmentHelper
 }
 
 export class LoadAssessmentButton extends React.Component<LoadAssessmentButtonProps> {
@@ -19,30 +21,30 @@ export class LoadAssessmentButton extends React.Component<LoadAssessmentButtonPr
         return (
             <InsightsCommandButton
                 iconProps={{ iconName: 'FabricOpenFolderHorizontal' }}
-                onClick={this.getAssessmentForUpload}
+                onClick={this.props.loadAssessmentHelper.getAssessmentForUpload}
             >
                 Load assessment
             </InsightsCommandButton>
         );
     }
 
-    private getAssessmentForUpload = () => {
-        const input = document.createElement('input');
-        input.type = 'file';
-        input.accept = '.a11ywebassessment';
-        input.onchange = e => {
-            const file = (e.target as HTMLInputElement).files[0];
-            const reader = new FileReader();
+    // private getAssessmentForUpload = () => {
+    //     const input = document.createElement('input');
+    //     input.type = 'file';
+    //     input.accept = '.a11ywebassessment';
+    //     input.onchange = e => {
+    //         const file = (e.target as HTMLInputElement).files[0];
+    //         const reader = new FileReader();
 
-            reader.onload = this.onReaderLoad;
-            reader.readAsText(file, 'UTF-8');
-        };
-        input.click();
-    };
+    //         reader.onload = this.onReaderLoad;
+    //         reader.readAsText(file, 'UTF-8');
+    //     };
+    //     input.click();
+    // };
 
-    private onReaderLoad = (readerEvent: ProgressEvent<FileReader>) => {
-        const content = readerEvent.target.result as string;
-        const assessmentData = this.props.deps.assessmentDataParser.parseAssessmentData(content);
-        this.props.deps.detailsViewActionMessageCreator.uploadAssessment(assessmentData);
-    };
+    // private onReaderLoad = (readerEvent: ProgressEvent<FileReader>) => {
+    //     const content = readerEvent.target.result as string;
+    //     const assessmentData = this.props.deps.assessmentDataParser.parseAssessmentData(content);
+    //     this.props.deps.detailsViewActionMessageCreator.uploadAssessment(assessmentData);
+    // };
 }
